refactor: share utils between resources adapters
